Added script that calls the linter
